+++ Source/gs/GSH_Vulkan/GSH_VulkanDraw.cpp
@@ -271,7 +271,7 @@ void CDraw::FlushVertices()
 	}
 
 	m_context->device.vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, drawPipeline->pipelineLayout,
-	                                          0, 1, &descriptorSet, descriptorDynamicOffsets.size(), descriptorDynamicOffsets.data());
+	                                          0, 1, &descriptorSet, static_cast<uint32_t>(descriptorDynamicOffsets.size()), descriptorDynamicOffsets.data());
 
 	m_context->device.vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, drawPipeline->pipeline);
 
@@ -409,7 +409,7 @@ VkDescriptorSet CDraw::PrepareDescriptorSet(VkDescriptorSetLayout descriptorSetL
 			}
 		}
 
-		m_context->device.vkUpdateDescriptorSets(m_context->device, writes.size(), writes.data(), 0, nullptr);
+		m_context->device.vkUpdateDescriptorSets(m_context->device, static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
 	}
 
 	m_descriptorSetCache.insert(std::make_pair(caps, descriptorSet));
